+++ frontend/src/app/page.tsx
@@ -1,6 +1,15 @@
 import { ChatInput } from "@/components/chat-input";
 import { fetcher } from "@/lib/fetcher";
 import { auth } from "@/lib/auth";
+import { getSettings } from "@/lib/cookies-server";
+import { redirect } from "next/navigation";
+
+async function isVlabAndProjectPopulated() {
+  const { projectID, virtualLabID } = await getSettings();
+  if (!(projectID && virtualLabID)) {
+    redirect("/settings");
+  }
+}
 
 async function getToolList() {
   const session = await auth();
@@ -18,6 +27,7 @@ async function getToolList() {
 }
 
 export default async function Home() {
+  await isVlabAndProjectPopulated();
   const availableTools = await getToolList();
   return (
     <div className="flex flex-col justify-center h-full">