+++ CHANGELOG.md
@@ -60,6 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - MarkDown rendering.
 - [frontend] - dark mode flickering.
 - [frontend] - Concatenate base url and endpoint correctly.
+- [frontend] - Better Vlab and Project selection.
 - [frontend] - Fix long message overflowing.
 - [frontent] - Scrolling while streaming.